Move skills to their own module
