+++ backend/belay_me/model/__init__.py
@@ -1,9 +1,11 @@
-from .user import (
-    User,
+from .skills import (
     Skills,
     YosemiteDecimalRating,
     VScaleRating,
-    create_user,
     create_rating,
     create_skills,
+)
+from .user import (
+    User,
+    create_user,
 )
